refactor: modularize main panels
